pyproject.toml
[tool.poetry]
name = "wakeword-training-record"
version = "0.1.0"
description = "Library to hande traning data for wakeword detectors."
authors = ["Robin-St <robin@farsk.net>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{include = "wakeword_training_record"}]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
iso639-lang = "^2.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
jsonschema = "^4.17.3"
flake8 = "^6.0.0"
black = "^22.12.0"
isort = "^5.12.0"
pre-commit = "^3.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
WTRtool = "wakeword_training_record.__main__:WTRtool"
